+++ SistemaGimnasio/Controllers/EntrenadorController.cs
@@ -6,10 +6,10 @@ namespace SistemaGimnasio.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class EntrenadorControllerRMB : ControllerBase
+    public class EntrenadorController : ControllerBase
     {
         private readonly IEntrenadorServiceRMB _service;
-        public EntrenadorControllerRMB(IEntrenadorServiceRMB service)
+        public EntrenadorController(IEntrenadorServiceRMB service)
         {
             _service = service;
         }
@@ -18,7 +18,7 @@ namespace SistemaGimnasio.Controllers
         public async Task<IActionResult> Post([FromBody] EntrenadorDtoRMB dto)
         {
             await _service.PostEntrenador(dto);
-            return Ok("El Entrenador fue registrado correctamente");
+            return Ok("El Entrenador fue registrado correctamente.");
         }
 
         [HttpGet("GetAllEntrenadores")]
@@ -31,8 +31,10 @@ namespace SistemaGimnasio.Controllers
         [HttpGet("GetEntrenadorById/{id}")]
         public async Task<IActionResult> GetById(int id)
         {
+            if (id <= 0) return BadRequest("El id no es valido.");
+
             var item = await _service.GetEntrenadorById(id);
-            if (item == null) return NotFound();
+            if (item == null) return NotFound("Entrenador no encontrado");
             return Ok(item);
         }
 
@@ -40,14 +42,16 @@ namespace SistemaGimnasio.Controllers
         public async Task<IActionResult> Update([FromBody] EntrenadorDtoRMB dto)
         {
             await _service.UpdateEntrenador(dto);
-            return Ok("Los datos del Entrenador fueron actualizados correctamente");
+            return Ok("Los datos del Entrenador fueron actualizados correctamente.");
         }
 
-        [HttpDelete("DeleteEntrenador{id}")]
+        [HttpDelete("DeleteEntrenador/{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (id <= 0) return BadRequest("El id no es valido.");
+
             await _service.DeleteEntrenador(id);
-            return Ok("Entrenador eliminado correctamente");
+            return Ok("Entrenador eliminado correctamente.");
         }
     }
 }
